requirements.txt
Flask==3.1.1
flask_sqlalchemy==3.1.1
python-dotenv==1.1.1
Requests==2.32.4
SQLAlchemy==2.0.43
